Fix linting issues in refactored code
- Fix int() wrapping of round() result in algorithms.py
- Split long lines in facade.py for BW parameters
- Replace ambiguous Unicode characters with ASCII equivalents
- Use more specific exception types in pillow_executor.py
- All ruff checks now pass

+++ src/iPhoto/core/filters/algorithms.py
@@ -212,9 +212,9 @@ def _clamp(value: float, min_val: float, max_val: float) -> float:
 def _float_to_uint8(value: float) -> int:
     """Convert *value* from ``[0.0, 1.0]`` to an 8-bit channel value."""
 
-    scaled = int(round(value * 255.0))
+    scaled = round(value * 255.0)
     if scaled < 0:
         return 0
     if scaled > 255:
         return 255
-    return scaled
+    return int(scaled)
